pyproject.toml
[project]
name = "techslop"
version = "0.1.0"
description = "Automated Tech News YouTube Shorts Pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0",
    "edge-tts>=6.1",
    "openai-whisper>=20231117",
    "feedparser>=6.0",
    "httpx>=0.27",
    "click>=8.1",
    "pydantic-settings>=2.0",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-api-python-client>=2.0",
    "pillow>=10.0",
    "python-dotenv>=1.0",
]

[project.scripts]
techslop = "pipeline:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
